refactor: extract validate_host_port to validate/common.rs
 deduplicates ~65 lines across endpoints.rs and listeners.rs.
 sanitize_identifier was already in common.rs.
