+++ crates/ffwd-config/src/validate/listeners.rs
@@ -1,8 +1,8 @@
 use crate::types::{ConfigError, InputType};
 use std::collections::HashMap;
 
-use super::common::{validation_error, validation_message};
-use super::endpoints::{canonical_listen_host_key, validate_host_port};
+use super::common::{validation_error, validation_message, validate_host_port};
+use super::endpoints::canonical_listen_host_key;
 use super::sensors::sensor_supported_families;
 
 /// Ensures a systemd unit name has a suffix, appending `.service` when absent.
