+++ crates/ffwd-config/src/validate/outputs.rs
@@ -3,8 +3,10 @@ use crate::types::{
 };
 use std::collections::HashMap;
 
-use super::common::{sanitize_identifier, validation_message};
-use super::endpoints::{validate_endpoint_url, validate_host_port};
+use super::common::sanitize_identifier;
+use super::common::validation_message;
+use super::endpoints::validate_endpoint_url;
+use super::common::validate_host_port;
 
 fn validate_url_output_endpoint(
     pipeline_name: &str,
